Fix: Constraints with invalid queries now get returned as failed constraint, not as if there were no issues found.
